requirements.txt
streamlit
requests
pandas
google-api-python-client
google-auth-oauthlib